12/27 기능수정 1. Tower의 상속관계 정리 2. TowerData의 상속관계 정리 3. UIManager에 TowerData, Tower Gameobject를 가지는 클래스를 만들고 이를  UIManager에 딕셔너리로 추가함. 타워들은 TowerType변수만 가지고 있으면 TowerData에 접근가능함. 4. 타워정보를 UI로 나타내는 곳에 전부 3번의 딕셔너리를 사용함. 5. TargetingTower가 공격 시 풀에서 꺼내는 Projectile도 딕셔너리를 통해 가지고 있을 수 있음.

+++ TowerDefense/Assets/Scripts/InterfaceControl/IHealable.cs
@@ -2,6 +2,6 @@ namespace InterfaceControl
 {
     public interface IHealable
     {
-        void Heal(in float amount);
+        void Heal(in ushort amount);
     }
 }
